+++ packages/gamut/src/Form/FormGroup.tsx
@@ -2,12 +2,16 @@ import styled from '@emotion/styled';
 import React, { HTMLAttributes } from 'react';
 
 import { Box } from '../Box';
+import { ToolTipProps } from '../ToolTip';
 import { FormError } from './FormError';
 import { FormGroupDescription } from './FormGroupDescription';
 import { FormGroupLabel } from './FormGroupLabel';
 
 export type FormGroupProps = HTMLAttributes<HTMLDivElement> & {
   label?: string;
+  /**
+   * [The for/id string of a label or labelable form-related element](https://developer.mozilla.org/en-US/docs/Web/API/HTMLLabelElement/htmlFor). The outer FormGroup or FormLabel should have an identical string as the inner FormElement for accessibility purposes.
+   */
   htmlFor?: string;
   className?: string;
   description?: string;
@@ -15,6 +19,7 @@ export type FormGroupProps = HTMLAttributes<HTMLDivElement> & {
   error?: string;
   disabled?: boolean;
   labelSize?: 'small' | 'large';
+  tooltip?: ToolTipProps;
 };
 
 const FormGroupContainer = styled(Box)<FormGroupProps>`
@@ -25,6 +30,7 @@ const FormGroupContainer = styled(Box)<FormGroupProps>`
 `;
 
 export const FormGroup: React.FC<FormGroupProps> = ({
+  tooltip,
   label,
   description,
   htmlFor,
@@ -42,6 +48,7 @@ export const FormGroup: React.FC<FormGroupProps> = ({
       showRequired={showRequired}
       size={labelSize}
       disabled={disabled}
+      tooltip={tooltip}
     >
       {label}
     </FormGroupLabel>