+++ packages/markdown-overrides/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.5.14](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.13...@codecademy/markdown-overrides@0.5.14) (2021-03-26)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.5.13](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.12...@codecademy/markdown-overrides@0.5.13) (2021-03-25)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.5.12](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.11...@codecademy/markdown-overrides@0.5.12) (2021-03-23)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
 ### [0.5.11](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.10...@codecademy/markdown-overrides@0.5.11) (2021-03-19)
 
 **Note:** Version bump only for package @codecademy/markdown-overrides